Use lazysizes for lazy loading and responsive image loading
Lazysizes will detect the best image to use depending on the screen size. This includes a few plugins to handle blurring the lower quality image placeholder, automatically handle aspect ratio for object-fit, and attribute changing (for when small thumbnails are clicked to replace the larger primary image). This change replaces the existing lazy load that used intersection observer.